Refactor: govkit structure and introduce marker management
- Added `cli/marker.py` to handle reading, writing, and migrating the .govkit marker.
- Introduced `cli/paths.py` for filesystem path management, centralizing directory references.
- Created `cli/version.py` to manage the package version as a single source of truth.
- Updated tests in `tests/test_maturity_model.py` and `tests/test_skill_context.py` to use the new paths module.
- Adjusted marker reading logic to support migration from legacy single-file format to directory structure.
- Implemented version comparison and migration warnings for backward compatibility.
