+++ tests/test_govkit.py
@@ -7,8 +7,18 @@ from pathlib import Path
 
 import pytest
 
-from cli.govkit import copy_entry, load_manifest, resolve_options, resolve_variant_files, cmd_apply, cmd_init, cmd_upgrade, write_govkit_marker, read_govkit_marker, _GOVKIT_VERSION
-
+from cli.govkit import (
+    cmd_apply,
+    cmd_init,
+    cmd_upgrade,
+    copy_entry,
+    load_manifest,
+    read_govkit_marker,
+    resolve_options,
+    resolve_variant_files,
+    write_govkit_marker,
+)
+from cli.version import GOVKIT_VERSION as _GOVKIT_VERSION
 
 # ---------------------------------------------------------------------------
 # Fixtures
@@ -101,37 +111,33 @@ class TestCopyEntry:
 
 class TestLoadManifest:
     def test_loads_valid_manifest(self, agent_dir, monkeypatch):
-        import cli.govkit as mod
 
-        monkeypatch.setattr(mod, "AGENTS_DIR", agent_dir / "agents")
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", agent_dir / "agents")
         m = load_manifest("test-agent")
         assert m["agent"] == "test-agent"
         assert "variants" in m
 
     def test_missing_agent_exits(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
-        monkeypatch.setattr(mod, "AGENTS_DIR", tmp_path)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", tmp_path)
         with pytest.raises(SystemExit):
             load_manifest("no-such-agent")
 
     def test_invalid_json_exits(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
         agent = tmp_path / "bad-agent"
         agent.mkdir()
         (agent / "manifest.json").write_text("not json!", encoding="utf-8")
-        monkeypatch.setattr(mod, "AGENTS_DIR", tmp_path)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", tmp_path)
         with pytest.raises(SystemExit):
             load_manifest("bad-agent")
 
     def test_missing_required_keys_exits(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
         agent = tmp_path / "incomplete"
         agent.mkdir()
         (agent / "manifest.json").write_text('{"agent": "x"}', encoding="utf-8")
-        monkeypatch.setattr(mod, "AGENTS_DIR", tmp_path)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", tmp_path)
         with pytest.raises(SystemExit):
             load_manifest("incomplete")
 
@@ -659,7 +665,7 @@ class TestMergeMode:
 
 class TestGovkitMarker:
     def test_write_and_read_marker(self, tmp_path):
-        from cli.govkit import write_govkit_marker, read_govkit_level
+        from cli.govkit import read_govkit_level, write_govkit_marker
 
         write_govkit_marker(tmp_path, "claude-code", "3", {"type": "api", "level": "3"})
         assert (tmp_path / ".govkit").exists()
@@ -679,6 +685,7 @@ class TestGovkitMarker:
 
     def test_marker_excludes_level_from_options(self, tmp_path):
         import json
+
         from cli.govkit import write_govkit_marker
 
         write_govkit_marker(tmp_path, "claude-code", "3", {"type": "api", "level": "3", "ci": "github"})
@@ -689,7 +696,8 @@ class TestGovkitMarker:
 
     def test_write_level_5_marker(self, tmp_path):
         import json
-        from cli.govkit import write_govkit_marker, read_govkit_level, _GOVKIT_VERSION
+
+        from cli.govkit import read_govkit_level, write_govkit_marker
 
         write_govkit_marker(tmp_path, "claude-code", "5", {"type": "api", "level": "5"})
         data = json.loads((tmp_path / ".govkit" / "marker.json").read_text(encoding="utf-8"))
@@ -725,7 +733,7 @@ class TestMarkerDirectoryLayout:
         assert not (tmp_path / ".govkit").is_file(), ".govkit must not be a single file"
 
     def test_read_new_directory_layout_returns_payload(self, tmp_path):
-        from cli.govkit import write_govkit_marker, read_govkit_marker
+        from cli.govkit import read_govkit_marker, write_govkit_marker
 
         write_govkit_marker(tmp_path, "claude-code", "4", {"type": "api", "ci": "github", "level": "4"})
 
@@ -880,7 +888,7 @@ class TestMarkerDirectoryLayout:
         assert (tmp_path / ".govkit" / "marker.json").is_file()
 
     def test_read_level_under_new_directory_layout(self, tmp_path):
-        from cli.govkit import write_govkit_marker, read_govkit_level
+        from cli.govkit import read_govkit_level, write_govkit_marker
 
         write_govkit_marker(tmp_path, "claude-code", "5", {"type": "api", "ci": "github", "level": "5"})
         assert read_govkit_level(tmp_path) == "5"
@@ -937,8 +945,9 @@ class TestMarkerMigrationDurability:
         file must still exist (or a recoverable backup must) AND the function
         must still return the parsed legacy data. The repo must never end up
         with no marker."""
-        import cli.govkit as mod
         from pathlib import Path
+
+        import cli.govkit as mod
         monkeypatch.setenv("GOVKIT_NO_DIRECTORY_MIGRATION_WARNING", "1")
         self._seed_legacy(tmp_path)
         legacy_path = tmp_path / ".govkit"
@@ -1002,6 +1011,7 @@ class TestExtendedMarkerFields:
 
     def test_new_marker_includes_empty_assumptions(self, tmp_path):
         import json
+
         from cli.govkit import write_govkit_marker
 
         write_govkit_marker(tmp_path, "claude-code", "4", {"type": "api", "ci": "github", "level": "4"})
@@ -1011,6 +1021,7 @@ class TestExtendedMarkerFields:
 
     def test_new_marker_includes_null_stack(self, tmp_path):
         import json
+
         from cli.govkit import write_govkit_marker
 
         write_govkit_marker(tmp_path, "claude-code", "4", {"type": "api", "ci": "github", "level": "4"})
@@ -1020,6 +1031,7 @@ class TestExtendedMarkerFields:
 
     def test_new_marker_includes_calibration_block(self, tmp_path):
         import json
+
         from cli.govkit import write_govkit_marker
 
         write_govkit_marker(tmp_path, "claude-code", "4", {"type": "api", "ci": "github", "level": "4"})
@@ -1031,6 +1043,7 @@ class TestExtendedMarkerFields:
     def test_write_marker_accepts_stack_param(self, tmp_path):
         """PR 2 will populate stack via this kwarg; PR 1 exposes the slot."""
         import json
+
         from cli.govkit import write_govkit_marker
 
         stack = {
@@ -1051,6 +1064,7 @@ class TestExtendedMarkerFields:
     def test_write_marker_accepts_assumptions_param(self, tmp_path):
         """PR 3 will populate assumptions via this kwarg; PR 1 exposes the slot."""
         import json
+
         from cli.govkit import write_govkit_marker
 
         assumptions = [{
@@ -1079,6 +1093,7 @@ class TestExtendedMarkerFields:
         read_govkit_marker must not crash; callers that need the fields use
         dict.get() defaults."""
         import json
+
         import cli.govkit as mod
         mod._reset_directory_migration_warning()
         mod._reset_migration_warning()
@@ -1135,6 +1150,7 @@ class TestIsUserEdited:
     def test_returns_false_when_mtime_at_or_before_applied_at(self, tmp_path):
         import os
         from datetime import datetime, timezone
+
         from cli.govkit import is_user_edited
 
         doc = tmp_path / "TECH_STACK.md"
@@ -1148,6 +1164,7 @@ class TestIsUserEdited:
     def test_returns_true_when_headed_file_modified_after_applied_at(self, tmp_path):
         import os
         from datetime import datetime, timezone
+
         from cli.govkit import is_user_edited
 
         doc = tmp_path / "TECH_STACK.md"
@@ -1190,6 +1207,7 @@ class TestCopyEntryEditProtection:
     def test_refuses_overwrite_of_user_edited_file(self, tmp_path, capsys):
         import os
         from datetime import datetime, timezone
+
         from cli.govkit import copy_entry
 
         src = tmp_path / "src.md"
@@ -1213,6 +1231,7 @@ class TestCopyEntryEditProtection:
     def test_force_overrides_protection(self, tmp_path, capsys):
         import os
         from datetime import datetime, timezone
+
         from cli.govkit import copy_entry
 
         src = tmp_path / "src.md"
@@ -1232,6 +1251,7 @@ class TestCopyEntryEditProtection:
     def test_copies_freely_when_file_unedited_since_apply(self, tmp_path):
         import os
         from datetime import datetime, timezone
+
         from cli.govkit import copy_entry
 
         src = tmp_path / "src.md"
@@ -1249,6 +1269,7 @@ class TestCopyEntryEditProtection:
     def test_directory_copy_protects_per_file(self, tmp_path):
         import os
         from datetime import datetime, timezone
+
         from cli.govkit import copy_entry
 
         src_dir = tmp_path / "src"
@@ -1329,6 +1350,7 @@ class TestL5OnlyGovernedExclusion:
 
     def test_l4_apply_excludes_llm_contracts_from_governed(self, tmp_path, monkeypatch):
         import argparse
+
         from cli.govkit import cmd_apply
 
         target = tmp_path / "project"
@@ -1352,6 +1374,7 @@ class TestL5OnlyGovernedExclusion:
 
     def test_l5_apply_includes_llm_contracts(self, tmp_path, monkeypatch):
         import argparse
+
         from cli.govkit import cmd_apply
 
         target = tmp_path / "project"
@@ -1381,6 +1404,7 @@ class TestApplyTypeDataStackDefault:
 
     def test_apply_type_data_defaults_to_dbt_stack(self, tmp_path):
         import argparse
+
         from cli.govkit import cmd_apply, read_govkit_marker
 
         target = tmp_path / "project"
@@ -1586,13 +1610,12 @@ def _apply_args(target: Path, **overrides) -> argparse.Namespace:
 
 class TestSmokeApply:
     def test_features_dir_created_empty(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
         fake_repo = _make_fake_agent(tmp_path)
         target = tmp_path / "target"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", fake_repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", fake_repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", fake_repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", fake_repo)
 
         cmd_apply(_apply_args(target))
 
@@ -1602,13 +1625,12 @@ class TestSmokeApply:
         assert starters == [], f"No starter_* folders should be copied; found: {starters}"
 
     def test_reminder_message_printed(self, tmp_path, monkeypatch, capsys):
-        import cli.govkit as mod
 
         fake_repo = _make_fake_agent(tmp_path)
         target = tmp_path / "target"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", fake_repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", fake_repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", fake_repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", fake_repo)
 
         cmd_apply(_apply_args(target))
 
@@ -1617,7 +1639,6 @@ class TestSmokeApply:
 
     def test_non_features_shared_still_copied(self, tmp_path, monkeypatch):
         """Shared entries that are not under features/ should still be copied."""
-        import cli.govkit as mod
 
         fake_repo = _make_fake_agent(tmp_path, shared=["features/starter_backend/", "governance/"])
         (fake_repo / "governance").mkdir()
@@ -1625,8 +1646,8 @@ class TestSmokeApply:
 
         target = tmp_path / "target"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", fake_repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", fake_repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", fake_repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", fake_repo)
 
         cmd_apply(_apply_args(target))
 
@@ -1635,13 +1656,12 @@ class TestSmokeApply:
 
     def test_apply_silent_when_no_extensions(self, tmp_path, monkeypatch, capsys):
         """Apply must not mention extensions when extensions/ is absent — regression guard."""
-        import cli.govkit as mod
 
         fake_repo = _make_fake_agent(tmp_path)
         target = tmp_path / "target"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", fake_repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", fake_repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", fake_repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", fake_repo)
 
         cmd_apply(_apply_args(target))
 
@@ -1649,7 +1669,6 @@ class TestSmokeApply:
 
     def test_apply_with_extensions_prints_summary(self, tmp_path, monkeypatch, capsys):
         """Apply should discover and report extensions present in the target."""
-        import cli.govkit as mod
 
         fake_repo = _make_fake_agent(tmp_path)
         target = tmp_path / "target"
@@ -1669,8 +1688,8 @@ class TestSmokeApply:
                 """),
             encoding="utf-8",
         )
-        monkeypatch.setattr(mod, "AGENTS_DIR", fake_repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", fake_repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", fake_repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", fake_repo)
 
         cmd_apply(_apply_args(target))
 
@@ -1689,10 +1708,9 @@ class TestSmokeInit:
         return tmp_path / "govkit_install"
 
     def test_init_scaffolds_from_bundled_starters(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
         fake_repo = self._bundled_repo(tmp_path)
-        monkeypatch.setattr(mod, "REPO_ROOT", fake_repo)
+        monkeypatch.setattr("cli.paths.REPO_ROOT", fake_repo)
 
         target = tmp_path / "project"
         (target / "features").mkdir(parents=True)
@@ -1707,10 +1725,9 @@ class TestSmokeInit:
 
     def test_init_ignores_starters_in_target(self, tmp_path, monkeypatch):
         """Starters left in the target project (from old govkit versions) should not be used."""
-        import cli.govkit as mod
 
         fake_repo = self._bundled_repo(tmp_path)
-        monkeypatch.setattr(mod, "REPO_ROOT", fake_repo)
+        monkeypatch.setattr("cli.paths.REPO_ROOT", fake_repo)
 
         target = tmp_path / "project"
         features_dir = target / "features"
@@ -1728,10 +1745,9 @@ class TestSmokeInit:
     @pytest.mark.parametrize("ui_starter", ["ui-react", "ui-angular"])
     def test_init_ui_variants_resolve_to_starter_ui(self, tmp_path, monkeypatch, ui_starter):
         """ui-react and ui-angular both map to the framework-agnostic starter_ui dir."""
-        import cli.govkit as mod
 
         fake_repo = self._bundled_repo(tmp_path, starter="starter_ui")
-        monkeypatch.setattr(mod, "REPO_ROOT", fake_repo)
+        monkeypatch.setattr("cli.paths.REPO_ROOT", fake_repo)
 
         target = tmp_path / "project"
         (target / "features").mkdir(parents=True)
@@ -1750,7 +1766,7 @@ class TestSmokeInit:
         """_resolve_starter_dir maps both UI variants to starter_ui (no starter_ui_l5 today)."""
         import cli.govkit as mod
         fake_repo = self._bundled_repo(tmp_path, starter="starter_ui")
-        monkeypatch.setattr(mod, "REPO_ROOT", fake_repo)
+        monkeypatch.setattr("cli.paths.REPO_ROOT", fake_repo)
 
         result = mod._resolve_starter_dir(ui_starter, level)
         assert result.name == "starter_ui", (
@@ -1891,13 +1907,12 @@ class TestCmdUpgrade:
 
     def test_upgrade_refreshes_agent_files(self, tmp_path, monkeypatch):
         """cmd_upgrade overwrites agent files (files category)."""
-        import cli.govkit as mod
 
         repo = _make_upgrade_repo(tmp_path)
         target = self._target_with_marker(tmp_path, repo, version="0.1.0")
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.2.0")
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.2.0")
 
         cmd_upgrade(argparse.Namespace(target=str(target), force=False))
 
@@ -1910,16 +1925,15 @@ class TestCmdUpgrade:
         header reflecting the new baseline; the file body equals the new
         baseline content.
         """
-        import cli.govkit as mod
 
         repo = _make_upgrade_repo(tmp_path)
         # Update the governed file in the repo to simulate a govkit update
         (repo / "docs" / "contract.md").write_text("# Contract v2", encoding="utf-8")
 
         target = self._target_with_marker(tmp_path, repo, version="0.1.0")
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.2.0")
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.2.0")
 
         cmd_upgrade(argparse.Namespace(target=str(target), force=False))
 
@@ -1930,12 +1944,11 @@ class TestCmdUpgrade:
 
     def test_upgrade_skips_when_version_current(self, tmp_path, monkeypatch, capsys):
         """cmd_upgrade does nothing when already at current version."""
-        import cli.govkit as mod
 
         repo = _make_upgrade_repo(tmp_path)
         target = self._target_with_marker(tmp_path, repo, version=_GOVKIT_VERSION)
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_upgrade(argparse.Namespace(target=str(target), force=False))
 
@@ -1946,12 +1959,11 @@ class TestCmdUpgrade:
 
     def test_upgrade_force_overwrites_even_if_current(self, tmp_path, monkeypatch):
         """cmd_upgrade --force overwrites even when version is current."""
-        import cli.govkit as mod
 
         repo = _make_upgrade_repo(tmp_path)
         target = self._target_with_marker(tmp_path, repo, version=_GOVKIT_VERSION)
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_upgrade(argparse.Namespace(target=str(target), force=True))
 
@@ -1959,13 +1971,12 @@ class TestCmdUpgrade:
 
     def test_upgrade_updates_marker_version(self, tmp_path, monkeypatch):
         """cmd_upgrade writes the new govkit version to the .govkit marker."""
-        import cli.govkit as mod
 
         repo = _make_upgrade_repo(tmp_path)
         target = self._target_with_marker(tmp_path, repo, version="0.1.0")
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.2.0")
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.2.0")
 
         cmd_upgrade(argparse.Namespace(target=str(target), force=False))
 
@@ -1974,11 +1985,10 @@ class TestCmdUpgrade:
 
     def test_upgrade_no_marker_exits(self, tmp_path, monkeypatch):
         """cmd_upgrade exits 1 when no .govkit marker exists."""
-        import cli.govkit as mod
 
         target = tmp_path / "bare"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", tmp_path / "agents")
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", tmp_path / "agents")
 
         with pytest.raises(SystemExit):
             cmd_upgrade(argparse.Namespace(target=str(target), force=False))
@@ -1992,9 +2002,10 @@ class TestCmdUpgradeEditProtection:
     def _setup_repo_and_target(self, tmp_path):
         """Build a tiny repo + target where the target has a headed,
         user-edited governed doc."""
-        import os
         import json
+        import os
         from datetime import datetime, timezone
+
         from cli.headers import format_editable_header
 
         repo = _make_upgrade_repo(tmp_path)
@@ -2026,12 +2037,11 @@ class TestCmdUpgradeEditProtection:
         return repo, target
 
     def test_upgrade_refuses_overwriting_edited_governed_file(self, tmp_path, monkeypatch, capsys):
-        import cli.govkit as mod
 
         repo, target = self._setup_repo_and_target(tmp_path)
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.2.0")
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.2.0")
 
         cmd_upgrade(argparse.Namespace(target=str(target), force=False))
 
@@ -2045,12 +2055,11 @@ class TestCmdUpgradeEditProtection:
         assert "--force" in err
 
     def test_upgrade_with_force_overwrites_edited_governed_file(self, tmp_path, monkeypatch, capsys):
-        import cli.govkit as mod
 
         repo, target = self._setup_repo_and_target(tmp_path)
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.2.0")
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.2.0")
 
         cmd_upgrade(argparse.Namespace(target=str(target), force=True))
 
@@ -2063,11 +2072,11 @@ class TestCmdUpgradeEditProtection:
         """A headed governed doc that hasn't been edited since applied_at
         gets refreshed normally, with the header regenerated for the new
         baseline."""
-        import os
         import json
+        import os
         from datetime import datetime, timezone
+
         from cli.headers import format_editable_header
-        import cli.govkit as mod
 
         repo = _make_upgrade_repo(tmp_path)
         (repo / "docs" / "contract.md").write_text("# Contract v2\n", encoding="utf-8")
@@ -2093,9 +2102,9 @@ class TestCmdUpgradeEditProtection:
         marker_dir.mkdir()
         (marker_dir / "marker.json").write_text(json.dumps(marker), encoding="utf-8")
 
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.2.0")
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.2.0")
 
         cmd_upgrade(argparse.Namespace(target=str(target), force=False))
 
@@ -2112,13 +2121,12 @@ class TestCmdApplyAddsEditableHeaders:
     def test_governed_md_files_get_header(self, tmp_path, monkeypatch):
         """When cmd_apply copies a governed .md doc, the doc gets the
         govkit:editable header prepended."""
-        import cli.govkit as mod
 
         repo = _make_upgrade_repo(tmp_path)
         target = tmp_path / "project"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_apply(argparse.Namespace(
             agent="test-agent",
@@ -2179,13 +2187,12 @@ class TestCmdApplyStackOverlay:
         return tmp_path
 
     def test_apply_with_default_stack_installs_python_fastapi(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
         repo = self._agent_with_stack(tmp_path)
         target = tmp_path / "project"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_apply(argparse.Namespace(
             agent="test-agent", target=str(target),
@@ -2198,13 +2205,12 @@ class TestCmdApplyStackOverlay:
         assert "baseline: python-fastapi@" in content
 
     def test_apply_with_explicit_stack_installs_overlay(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
         repo = self._agent_with_stack(tmp_path)
         target = tmp_path / "project"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_apply(argparse.Namespace(
             agent="test-agent", target=str(target),
@@ -2218,13 +2224,12 @@ class TestCmdApplyStackOverlay:
         assert "C#" in content or ".NET" in content
 
     def test_marker_records_stack_metadata(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
         repo = self._agent_with_stack(tmp_path)
         target = tmp_path / "project"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_apply(argparse.Namespace(
             agent="test-agent", target=str(target),
@@ -2239,13 +2244,12 @@ class TestCmdApplyStackOverlay:
         assert marker["options"]["stack"] == "dotnet-aspnet"
 
     def test_marker_records_stack_assumption_source_flag_when_cli_provided(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
         repo = self._agent_with_stack(tmp_path)
         target = tmp_path / "project"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_apply(argparse.Namespace(
             agent="test-agent", target=str(target),
@@ -2259,13 +2263,12 @@ class TestCmdApplyStackOverlay:
         assert stack_assumption["source"] == "flag"
 
     def test_marker_records_stack_assumption_source_default_when_no_flag(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
         repo = self._agent_with_stack(tmp_path)
         target = tmp_path / "project"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_apply(argparse.Namespace(
             agent="test-agent", target=str(target),
@@ -2281,7 +2284,6 @@ class TestCmdApplyStackOverlay:
     def test_ui_type_does_not_apply_stack_overlay(self, tmp_path, monkeypatch):
         """UI installs target docs/ui/architecture/, not docs/backend/. Stack
         overlays only ship backend docs, so they must be a no-op for UI types."""
-        import cli.govkit as mod
 
         agent = "ui-test-agent"
         agents = tmp_path / "agents" / agent
@@ -2315,8 +2317,8 @@ class TestCmdApplyStackOverlay:
 
         target = tmp_path / "project"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", tmp_path / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", tmp_path)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", tmp_path / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", tmp_path)
 
         cmd_apply(argparse.Namespace(
             agent=agent, target=str(target),
@@ -2367,7 +2369,6 @@ class TestCmdApplyDetection:
         return tmp_path
 
     def test_dotnet_target_with_no_stack_flag_infers_dotnet_aspnet(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
         repo = self._agent_with_stack(tmp_path)
         target = tmp_path / "project"
@@ -2379,8 +2380,8 @@ class TestCmdApplyDetection:
         )
         (target / "global.json").write_text('{"sdk":{}}', encoding="utf-8")
 
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_apply(argparse.Namespace(
             agent="test-agent", target=str(target),
@@ -2399,7 +2400,6 @@ class TestCmdApplyDetection:
     def test_explicit_stack_flag_overrides_detection(self, tmp_path, monkeypatch):
         """User intent (--stack) always wins. The marker reflects the explicit
         choice with source='flag', not the inferred stack."""
-        import cli.govkit as mod
 
         repo = self._agent_with_stack(tmp_path)
         target = tmp_path / "project"
@@ -2409,8 +2409,8 @@ class TestCmdApplyDetection:
             '<Project Sdk="Microsoft.NET.Sdk.Web"></Project>\n', encoding="utf-8",
         )
 
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_apply(argparse.Namespace(
             agent="test-agent", target=str(target),
@@ -2425,15 +2425,14 @@ class TestCmdApplyDetection:
     def test_empty_target_falls_back_to_default(self, tmp_path, monkeypatch):
         """No detectable signals → use python-fastapi as the default and
         flag it review_required so the team knows we're guessing."""
-        import cli.govkit as mod
 
         repo = self._agent_with_stack(tmp_path)
         target = tmp_path / "project"
         target.mkdir()
         # No signals at all in target
 
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_apply(argparse.Namespace(
             agent="test-agent", target=str(target),
@@ -2448,7 +2447,6 @@ class TestCmdApplyDetection:
 
     def test_apply_prints_detection_summary(self, tmp_path, monkeypatch, capsys):
         """cmd_apply emits a 'detecting repo profile' block before installing."""
-        import cli.govkit as mod
 
         repo = self._agent_with_stack(tmp_path)
         target = tmp_path / "project"
@@ -2458,8 +2456,8 @@ class TestCmdApplyDetection:
         )
         (target / "global.json").write_text('{"sdk":{}}', encoding="utf-8")
 
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_apply(argparse.Namespace(
             agent="test-agent", target=str(target),
@@ -2510,7 +2508,6 @@ class TestCmdApplyDetectFlag:
         return tmp_path
 
     def test_detect_flag_writes_nothing_to_target(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
         repo = self._agent(tmp_path)
         target = tmp_path / "project"
@@ -2519,8 +2516,8 @@ class TestCmdApplyDetectFlag:
         (target / "Api.csproj").write_text(
             '<Project Sdk="Microsoft.NET.Sdk.Web"></Project>\n', encoding="utf-8",
         )
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_apply(argparse.Namespace(
             agent="test-agent", target=str(target),
@@ -2539,7 +2536,6 @@ class TestCmdApplyDetectFlag:
         assert (target / "Api.csproj").is_file()
 
     def test_detect_flag_prints_proposed_config(self, tmp_path, monkeypatch, capsys):
-        import cli.govkit as mod
 
         repo = self._agent(tmp_path)
         target = tmp_path / "project"
@@ -2548,8 +2544,8 @@ class TestCmdApplyDetectFlag:
             '<Project Sdk="Microsoft.NET.Sdk.Web"></Project>\n', encoding="utf-8",
         )
         (target / "global.json").write_text('{"sdk":{}}', encoding="utf-8")
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_apply(argparse.Namespace(
             agent="test-agent", target=str(target),
@@ -2566,13 +2562,12 @@ class TestCmdApplyDetectFlag:
         assert "dry-run" in out.lower() or "--detect" in out
 
     def test_detect_flag_on_empty_target_still_exits_cleanly(self, tmp_path, monkeypatch, capsys):
-        import cli.govkit as mod
 
         repo = self._agent(tmp_path)
         target = tmp_path / "project"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         # Should not raise
         cmd_apply(argparse.Namespace(
@@ -2586,7 +2581,6 @@ class TestCmdApplyDetectFlag:
         """Mirror of TestResolveStackChoiceTypeCompatibility, applied to the
         --detect dry-run path. The reported stack must match what real apply
         would pick — type-incompatible inferences are rejected in both places."""
-        import cli.govkit as mod
 
         repo = self._agent(tmp_path)
         target = tmp_path / "project"
@@ -2596,8 +2590,8 @@ class TestCmdApplyDetectFlag:
             '[project]\nname = "demo"\ndependencies = ["fastapi>=0.100"]\n',
             encoding="utf-8",
         )
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_apply(argparse.Namespace(
             agent="test-agent", target=str(target),
@@ -2647,13 +2641,12 @@ class TestCmdStackApply:
     def _existing_install(self, tmp_path, monkeypatch, stack_id="python-fastapi"):
         """Build a target with an existing .govkit/marker.json (as written by
         cmd_apply) so cmd_stack_apply has something to read."""
-        import cli.govkit as mod
 
         repo = _make_upgrade_repo(tmp_path)
         target = tmp_path / "project"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
         cmd_apply(argparse.Namespace(
             agent="test-agent", target=str(target),
             level="4", type="api", ci="github", stack=stack_id, force=False,
@@ -2713,6 +2706,7 @@ class TestCmdStackApply:
         stack apply without --force."""
         import os
         from datetime import datetime, timezone
+
         from cli.govkit import cmd_stack_apply
 
         target = self._existing_install(tmp_path, monkeypatch, stack_id="python-fastapi")
@@ -2741,13 +2735,12 @@ class TestCmdApplyWritesSetupReview:
     prints the Review Checklist banner."""
 
     def test_apply_writes_setup_review_file(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
         repo = _make_upgrade_repo(tmp_path)
         target = tmp_path / "project"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_apply(argparse.Namespace(
             agent="test-agent", target=str(target),
@@ -2761,13 +2754,12 @@ class TestCmdApplyWritesSetupReview:
         assert "test-agent" in content
 
     def test_apply_prints_review_checklist_banner(self, tmp_path, monkeypatch, capsys):
-        import cli.govkit as mod
 
         repo = _make_upgrade_repo(tmp_path)
         target = tmp_path / "project"
         target.mkdir()
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
 
         cmd_apply(argparse.Namespace(
             agent="test-agent", target=str(target),
@@ -2782,7 +2774,7 @@ class TestCmdApplyWritesSetupReview:
         """cmd_upgrade must also (re)write GOVKIT_SETUP_REVIEW.md so the
         review reflects the post-upgrade state, not the prior install."""
         import json
-        import cli.govkit as mod
+
 
         repo = _make_upgrade_repo(tmp_path)
         target = tmp_path / "project"
@@ -2799,9 +2791,9 @@ class TestCmdApplyWritesSetupReview:
         # Pre-existing agent file so upgrade has something to refresh:
         (target / "CLAUDE.md").write_text("# old\n", encoding="utf-8")
 
-        monkeypatch.setattr(mod, "AGENTS_DIR", repo / "agents")
-        monkeypatch.setattr(mod, "REPO_ROOT", repo)
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.2.0")
+        monkeypatch.setattr("cli.paths.AGENTS_DIR", repo / "agents")
+        monkeypatch.setattr("cli.paths.REPO_ROOT", repo)
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.2.0")
 
         cmd_upgrade(argparse.Namespace(target=str(target), force=False))
 
@@ -2923,7 +2915,7 @@ class TestNoUiDimensionInManifests:
 
     @pytest.mark.parametrize("agent", ["claude-code", "codex", "copilot"])
     def test_no_ui_options_or_variants(self, agent):
-        from cli.govkit import AGENTS_DIR
+        from cli.paths import AGENTS_DIR
         manifest_path = AGENTS_DIR / agent / "manifest.json"
         manifest = json.loads(manifest_path.read_text(encoding="utf-8"))
         assert "ui" not in manifest.get("options", {}), (
@@ -2935,7 +2927,7 @@ class TestNoUiDimensionInManifests:
 
     @pytest.mark.parametrize("agent", ["claude-code", "codex", "copilot"])
     def test_type_choices_include_ui_shapes(self, agent):
-        from cli.govkit import AGENTS_DIR
+        from cli.paths import AGENTS_DIR
         manifest_path = AGENTS_DIR / agent / "manifest.json"
         manifest = json.loads(manifest_path.read_text(encoding="utf-8"))
         type_choices = manifest["options"]["type"]["choices"]
@@ -3034,9 +3026,8 @@ class TestUpgradeMigrateLevels:
         mod._reset_migration_warning()
 
     def test_already_at_070_is_noop(self, tmp_path, monkeypatch, capsys):
-        import cli.govkit as mod
 
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.7.0")
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.7.0")
         target = _make_legacy_target(tmp_path, level="3", version="0.7.0")
 
         with pytest.raises(SystemExit) as exc_info:
@@ -3048,9 +3039,8 @@ class TestUpgradeMigrateLevels:
         assert read_govkit_marker(target)["version"] == "0.7.0"
 
     def test_l4_v06_bumps_version_only(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.7.0")
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.7.0")
         target = _make_legacy_target(tmp_path, level="4", version="0.6.0", three_artifact=False)
 
         with pytest.raises(SystemExit) as exc_info:
@@ -3063,9 +3053,8 @@ class TestUpgradeMigrateLevels:
         assert "level" not in marker["options"]
 
     def test_l5_v06_bumps_version_only(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.7.0")
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.7.0")
         target = _make_legacy_target(tmp_path, level="5", version="0.6.0", three_artifact=False)
 
         with pytest.raises(SystemExit) as exc_info:
@@ -3076,9 +3065,8 @@ class TestUpgradeMigrateLevels:
         assert marker["level"] == "5"
 
     def test_l3_v06_choice_1_generates_stubs_and_migrates_to_l4(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.7.0")
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.7.0")
         monkeypatch.setattr("builtins.input", lambda _: "1")
         target = _make_legacy_target(tmp_path, level="3", version="0.6.0", num_features=2)
 
@@ -3096,10 +3084,8 @@ class TestUpgradeMigrateLevels:
 
     def test_l3_v06_choice_1_stub_eval_criteria_fails_validation(self, tmp_path, monkeypatch):
         """The generated eval_criteria.yaml stub must fail validation (TBD placeholders)."""
-        import cli.govkit as mod
-        from cli.validate import check_eval_criteria
 
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.7.0")
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.7.0")
         monkeypatch.setattr("builtins.input", lambda _: "1")
         target = _make_legacy_target(tmp_path, level="3", version="0.6.0", num_features=1)
 
@@ -3116,9 +3102,8 @@ class TestUpgradeMigrateLevels:
         assert "mode: TBD" in text
 
     def test_l3_v06_choice_1_stub_preflight_has_tbd_sections(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.7.0")
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.7.0")
         monkeypatch.setattr("builtins.input", lambda _: "1")
         target = _make_legacy_target(tmp_path, level="3", version="0.6.0", num_features=1)
 
@@ -3133,9 +3118,8 @@ class TestUpgradeMigrateLevels:
         assert text.count("TBD") >= 9, "stub should have TBD per section"
 
     def test_l3_v06_choice_1_does_not_overwrite_existing_artifacts(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.7.0")
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.7.0")
         monkeypatch.setattr("builtins.input", lambda _: "1")
         target = _make_legacy_target(tmp_path, level="3", version="0.6.0", num_features=1)
         # Pre-existing eval_criteria.yaml in one feature
@@ -3149,9 +3133,8 @@ class TestUpgradeMigrateLevels:
         assert existing.read_text(encoding="utf-8") == "version: '1'\nmode: deterministic\n"
 
     def test_l3_v06_choice_2_no_stubs_only_marker_rewrite(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.7.0")
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.7.0")
         monkeypatch.setattr("builtins.input", lambda _: "2")
         target = _make_legacy_target(tmp_path, level="3", version="0.6.0", num_features=2)
 
@@ -3167,9 +3150,8 @@ class TestUpgradeMigrateLevels:
             assert not (fd / "architecture_preflight.md").exists()
 
     def test_l3_v06_choice_3_confirmed_deletes_features(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.7.0")
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.7.0")
         responses = iter(["3", "yes"])
         monkeypatch.setattr("builtins.input", lambda _: next(responses))
         target = _make_legacy_target(tmp_path, level="3", version="0.6.0", num_features=2)
@@ -3184,9 +3166,8 @@ class TestUpgradeMigrateLevels:
         assert not (target / "features").exists()
 
     def test_l3_v06_choice_3_unconfirmed_keeps_features(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.7.0")
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.7.0")
         responses = iter(["3", "no"])
         monkeypatch.setattr("builtins.input", lambda _: next(responses))
         target = _make_legacy_target(tmp_path, level="3", version="0.6.0", num_features=2)
@@ -3201,9 +3182,8 @@ class TestUpgradeMigrateLevels:
         assert marker["version"] == "0.6.0"
 
     def test_l3_v06_choice_4_aborts_no_changes(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.7.0")
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.7.0")
         monkeypatch.setattr("builtins.input", lambda _: "4")
         target = _make_legacy_target(tmp_path, level="3", version="0.6.0", num_features=2)
 
@@ -3215,9 +3195,8 @@ class TestUpgradeMigrateLevels:
         assert marker["version"] == "0.6.0"
 
     def test_l3_v06_invalid_choice_aborts(self, tmp_path, monkeypatch):
-        import cli.govkit as mod
 
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.7.0")
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.7.0")
         monkeypatch.setattr("builtins.input", lambda _: "9")
         target = _make_legacy_target(tmp_path, level="3", version="0.6.0", num_features=1)
 
@@ -3232,7 +3211,7 @@ class TestUpgradeMigrateLevels:
         """After --migrate-levels rewrites the marker to 0.7.0, the warning stops firing."""
         import cli.govkit as mod
 
-        monkeypatch.setattr(mod, "_GOVKIT_VERSION", "0.7.0")
+        monkeypatch.setattr("cli.version.GOVKIT_VERSION", "0.7.0")
         monkeypatch.setattr("builtins.input", lambda _: "2")
         monkeypatch.delenv("GOVKIT_NO_MIGRATION_WARNING", raising=False)
         target = _make_legacy_target(tmp_path, level="3", version="0.6.0", num_features=1)
